flytter ut ikoner i egen fil
